move models to submodule, move authzClient example to own submodule
