+++ models/pom.xml
@@ -1,0 +1,125 @@
+<?xml version="1.0"?>
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+         xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>io.xacml</groupId>
+        <artifactId>json-pep-sdk</artifactId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
+    <artifactId>json-pep-sdk-models</artifactId>
+    <version>1.0-SNAPSHOT</version>
+    <name>json-pep-sdk-models</name>
+    <url>http://maven.apache.org</url>
+    <packaging>jar</packaging>
+    <dependencies>
+        <dependency>
+            <groupId>commons-io</groupId>
+            <artifactId>commons-io</artifactId>
+            <version>2.6</version>
+        </dependency>
+        <dependency>
+            <groupId>io.springfox</groupId>
+            <artifactId>springfox-swagger2</artifactId>
+            <version>2.9.2</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-databind</artifactId>
+            <version>${jackson.version}</version>
+        </dependency>
+    </dependencies>
+    <build>
+        <plugins>
+            <!--<plugin>-->
+                <!--<groupId>com.github.kongchen</groupId>-->
+                <!--<artifactId>swagger-maven-plugin</artifactId>-->
+                <!--<version>${swagger-maven-plugin-version}</version>-->
+                <!--<configuration>-->
+                    <!--<apiSources>-->
+                        <!--<apiSource>-->
+                            <!--<springmvc>false</springmvc>-->
+                            <!--<locations>-->
+                                <!--<location>io.xacml.json.model</location>-->
+                            <!--</locations>-->
+                            <!--<schemes>-->
+                                <!--<scheme>http</scheme>-->
+                                <!--<scheme>https</scheme>-->
+                            <!--</schemes>-->
+                            <!--&lt;!&ndash;<host>www.example.com:8080</host>&ndash;&gt;-->
+                            <!--&lt;!&ndash;<basePath>/api</basePath>&ndash;&gt;-->
+                            <!--<info>-->
+                                <!--<title>Axiomatics REST endpoint</title>-->
+                                <!--<version>v1</version>-->
+                                <!--&lt;!&ndash; use markdown here because I'm using markdown for output,-->
+                                <!--if you need to use html or other markup language, you need to use your target language,-->
+                                 <!--and note escape your description for xml &ndash;&gt;-->
+                                <!--<description>-->
+                                    <!--This is a sample.-->
+                                <!--</description>-->
+                                <!--<termsOfService>-->
+                                    <!--http://www.github.com/kongchen/swagger-maven-plugin-->
+                                <!--</termsOfService>-->
+                                <!--<contact>-->
+                                    <!--<email>kongchen@gmail.com</email>-->
+                                    <!--<name>Kong Chen</name>-->
+                                    <!--<url>http://kongch.com</url>-->
+                                <!--</contact>-->
+                                <!--<license>-->
+                                    <!--<url>http://www.apache.org/licenses/LICENSE-2.0.html</url>-->
+                                    <!--<name>Apache 2.0</name>-->
+                                <!--</license>-->
+                            <!--</info>-->
+                            <!--<securityDefinitions>-->
+                                <!--<securityDefinition>-->
+                                    <!--<name>basicAuth</name>-->
+                                    <!--<type>basic</type>-->
+                                <!--</securityDefinition>-->
+                                <!--<securityDefinition>-->
+                                    <!--<json>/securityDefinition.json</json>-->
+                                <!--</securityDefinition>-->
+                            <!--</securityDefinitions>-->
+                            <!--&lt;!&ndash; Support classpath or file absolute path here.-->
+                            <!--1) classpath e.g: "classpath:/markdown.hbs", "classpath:/templates/hello.html"-->
+                            <!--2) file e.g: "${basedir}/src/main/resources/markdown.hbs",-->
+                                <!--"${basedir}/src/main/resources/template/hello.html" &ndash;&gt;-->
+                            <!--<templatePath>${basedir}/src/test/resources/strapdown.html.hbs</templatePath>-->
+                            <!--<outputPath>${basedir}/generated/document.html</outputPath>-->
+                            <!--<swaggerDirectory>${basedir}/generated/swagger-ui</swaggerDirectory>-->
+                            <!--<swaggerApiReader>com.wordnik.swagger.jaxrs.reader.DefaultJaxrsApiReader</swaggerApiReader>-->
+                            <!--<attachSwaggerArtifact>true</attachSwaggerArtifact>-->
+                            <!--<modelConverters>io.swagger.validator.BeanValidator</modelConverters>-->
+                            <!--<swaggerExtensions>-->
+                                <!--<swaggerExtension>com.example.VendorExtension</swaggerExtension>-->
+                            <!--</swaggerExtensions>-->
+                            <!--<enabledObjectMapperFeatures>-->
+                                <!--<feature>com.fasterxml.jackson.databind.SerializationFeature.WRITE_ENUMS_USING_TO_STRING</feature>-->
+                                <!--<feature>com.fasterxml.jackson.core.JsonParser$Feature.ALLOW_NUMERIC_LEADING_ZEROS</feature>-->
+                            <!--</enabledObjectMapperFeatures>-->
+                            <!--<disabledObjectMapperFeatures>-->
+                                <!--<feature>com.fasterxml.jackson.databind.SerializationFeature.FAIL_ON_EMPTY_BEANS</feature>-->
+                            <!--</disabledObjectMapperFeatures>-->
+                            <!--<operationIdFormat>{{className}}_{{methodName}}_{{httpMethod}}</operationIdFormat>-->
+                            <!--<externalDocs>-->
+                                <!--<description>Example external docs</description>-->
+                                <!--<url>https://example.com/docs</url>-->
+                            <!--</externalDocs>-->
+                        <!--</apiSource>-->
+                    <!--</apiSources>-->
+                <!--</configuration>-->
+                <!--<executions>-->
+                    <!--<execution>-->
+                        <!--<phase>compile</phase>-->
+                        <!--<goals>-->
+                            <!--<goal>generate</goal>-->
+                        <!--</goals>-->
+                    <!--</execution>-->
+                <!--</executions>-->
+            <!--</plugin>-->
+        </plugins>
+    </build>
+    <description>This project provides a sample PEP SDK for Java that generates a XACML request and response in
+        accordance with the JSON Profile of XACML WD22.
+    </description>
+</project>
