+++ models/src/main/java/io/xacml/json/model/Category.java
