+++ models/src/main/java/io/xacml/json/model/NamespaceDeclaration.java
