+++ models/src/main/java/io/xacml/json/model/Result.java
