+++ models/src/main/java/io/xacml/json/model/SingleResponse.java
