+++ models/src/test/java/io/xacml/json/model/JsonTest.java
@@ -3,17 +3,6 @@ package io.xacml.json.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-
-import io.xacml.json.model.Category;
-import io.xacml.json.model.MultiRequests;
-import io.xacml.json.model.PDPDecision;
-import io.xacml.json.model.Request;
-import io.xacml.json.model.Response;
-import io.xacml.json.model.Result;
-import io.xacml.json.model.Status;
-import io.xacml.json.model.StatusCode;
-import io.xacml.json.model.XPathExpression;
-
 import org.junit.Before;
 import org.junit.Test;
 
@@ -25,6 +14,7 @@ import java.util.List;
 import static org.junit.Assert.assertEquals;
 
 /**
+ *
  */
 public class JsonTest {
