+++ models/src/test/resources/json-examples/attribute-integer-inferred.json
