+++ models/src/test/resources/json-examples/attribute-list-string.json
