+++ models/src/test/resources/json-examples/attribute-xpath.json
