+++ models/src/test/resources/json-examples/multirequests.json
@@ -1,0 +1,22 @@
+{
+  "RequestReference": [
+    {
+      "ReferenceId": [
+        "foo1",
+        "bar1"
+      ]
+    },
+    {
+      "ReferenceId": [
+        "foo2",
+        "bar1"
+      ]
+    },
+    {
+      "ReferenceId": [
+        "foo3",
+        "bar1"
+      ]
+    }
+  ]
+}
