+++ models/src/test/resources/json-examples/request-with-base64-encoded-content.json
