+++ models/src/test/resources/json-examples/request-with-category.json
