+++ models/src/test/resources/json-examples/response-with-status-code.json
