+++ pom.xml
@@ -1,104 +1,85 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<groupId>io.xacml</groupId>
-	<artifactId>java-json-pep-sdk</artifactId>
-	<packaging>jar</packaging>
-	<version>1.0-SNAPSHOT</version>
-	<name>java-json-pep-sdk</name>
-	<url>http://maven.apache.org</url>
-	<dependencies>
-		<dependency>
-			<groupId>commons-io</groupId>
-			<artifactId>commons-io</artifactId>
-			<version>2.6</version>
-		</dependency>
-
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>${junit.version}</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>io.springfox</groupId>
-			<artifactId>springfox-swagger2</artifactId>
-			<version>2.9.2</version>
-		</dependency>
-		<dependency>
-			<groupId>com.fasterxml.jackson.core</groupId>
-			<artifactId>jackson-databind</artifactId>
-			<version>${jackson.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.glassfish.jersey.core</groupId>
-			<artifactId>jersey-client</artifactId>
-			<version>2.27</version>
-		</dependency>
-		<dependency>
-			<groupId>org.glassfish.jersey.inject</groupId>
-			<artifactId>jersey-hk2</artifactId>
-			<version>2.27</version>
-		</dependency>
-		<dependency>
-			<groupId>org.glassfish.jersey.media</groupId>
-			<artifactId>jersey-media-json-jackson</artifactId>
-			<version>2.27</version>
-		</dependency>
-		<dependency>
-			<groupId>org.projectlombok</groupId>
-			<artifactId>lombok</artifactId>
-			<version>${lombok.version}</version>
-			<scope>provided</scope>
-		</dependency>
-	</dependencies>
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-source-plugin</artifactId>
-				<executions>
-					<execution>
-						<id>attach-sources</id>
-						<goals>
-							<goal>jar</goal>
-						</goals>
-					</execution>
-				</executions>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-javadoc-plugin</artifactId>
-				<version>${maven-javadoc-plugin.version}</version>
-				<configuration>
-					<show>private</show>
-					<nohelp>true</nohelp>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
-	<description>This project provides a sample PEP SDK for Java that generates a XACML request and response in
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <groupId>io.xacml</groupId>
+    <artifactId>json-pep-sdk</artifactId>
+    <packaging>pom</packaging>
+    <version>1.0-SNAPSHOT</version>
+    <name>json-pep-sdk</name>
+    <url>http://maven.apache.org</url>
+    <modules>
+        <module>examples</module>
+        <module>models</module>
+    </modules>
+    <dependencies>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>${junit.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>${lombok.version}</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+            <version>1.7.21</version>
+        </dependency>
+    </dependencies>
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-source-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <id>attach-sources</id>
+                        <goals>
+                            <goal>jar</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-javadoc-plugin</artifactId>
+                <version>${maven-javadoc-plugin.version}</version>
+                <configuration>
+                    <show>private</show>
+                    <nohelp>true</nohelp>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+    <description>This project provides a sample PEP SDK for Java that generates a XACML request and response in
         accordance with the JSON Profile of XACML WD22.
     </description>
-	<organization>
-		<name>Axiomatics</name>
-		<url>https://www.axiomatics.com</url>
-	</organization>
-	<scm>
-		<url>https://github.com/davidjbrossard/java-json-pep-sdk</url>
-		<connection>scm:git:git://github.com/davidjbrossard/java-json-pep-sdk.git</connection>
-		<developerConnection>scm:git:git@github.com/davidjbrossard/java-json-pep-sdk.git</developerConnection>
-	</scm>
-	<properties>
-		<!-- marshalling -->
-		<jackson.version>[2.9.8,)</jackson.version>
-		<jersey.version>2.27</jersey.version>
-		<jersey-h2.version>2.27</jersey-h2.version>
-		<junit.version>4.12</junit.version>
-		<lombok.version>1.18.2</lombok.version>
-		<maven-javadoc-plugin.version>3.0.1</maven-javadoc-plugin.version>
-		<maven.compiler.target>1.8</maven.compiler.target>
-		<maven.compiler.source>1.8</maven.compiler.source>
-	</properties>
+    <organization>
+        <name>Axiomatics</name>
+        <url>https://www.axiomatics.com</url>
+    </organization>
+    <scm>
+        <url>https://github.com/davidjbrossard/java-json-pep-sdk</url>
+        <connection>scm:git:git://github.com/davidjbrossard/java-json-pep-sdk.git</connection>
+        <developerConnection>scm:git:git@github.com/davidjbrossard/java-json-pep-sdk.git</developerConnection>
+    </scm>
+    <properties>
+        <!-- marshalling -->
+        <kotlin.version>1.3.20</kotlin.version>
+        <jackson.version>[2.9.8,)</jackson.version>
+        <jersey.version>2.27</jersey.version>
+        <feign.version>9.3.1</feign.version>
+        <jersey-h2.version>2.27</jersey-h2.version>
+        <junit.version>4.12</junit.version>
+        <lombok.version>1.18.2</lombok.version>
+        <!--<swagger-maven-plugin-versio    n>3.1.8</swagger-maven-plugin-version>-->
+        <maven-javadoc-plugin.version>3.0.1</maven-javadoc-plugin.version>
+        <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.source>1.8</maven.compiler.source>
+    </properties>
 </project>
