--- src/test/resources/json-examples/multirequests.json
@@ -1,24 +1,0 @@
-{
-  "MultiRequests": {
-    "RequestReference": [
-      {
-        "ReferenceId": [
-          "foo1",
-          "bar1"
-        ]
-      },
-      {
-        "ReferenceId": [
-          "foo2",
-          "bar1"
-        ]
-      },
-      {
-        "ReferenceId": [
-          "foo3",
-          "bar1"
-        ]
-      }
-    ]
-  }
-}
